requirements.txt
streamlit
duckduckgo_search
trafilatura
beautifulsoup4
requests
pandas
openpyxl
python-dotenv
openai
google-genai
curl_cffi
pytest
pytest-mock
